Split ui into interface and interaction
